extract bio, about model fields into mixin
