+++ app/controllers/users_controller.rb
@@ -238,7 +238,7 @@ class UsersController < ApplicationController
       user.waifu = changes[:waifu]
       user.website = changes[:website]
       user.waifu_or_husbando = changes[:waifu_or_husbando]
-      user.bio = changes[:mini_bio] || ""
+      user.bio = changes[:bio] || ""
       user.waifu_char_id = changes[:waifu_char_id]
 
       if Rails.env.production? and changes[:cover_image_url] =~ /^data:image/
