+++ app/frontend/app/models/group.js
@@ -1,6 +1,7 @@
 import DS from 'ember-data';
+import ModelTruncatedDetails from '../mixins/model-truncated-details';
 
-export default DS.Model.extend({
+export default DS.Model.extend(ModelTruncatedDetails, {
   name: DS.attr('string'),
   bio: DS.attr('string'),
   about: DS.attr('string'),
@@ -23,25 +24,5 @@ export default DS.Model.extend({
   isMember: function(user) {
     return this.get('members').filterBy('pending', false)
       .findBy('user.id', user.get('id'));
-  },
-
-  truncatedBio: function(key, value) {
-    // setter
-    if (arguments.length > 1) {
-      this.set('bio', value.slice(0, 140));
-    }
-
-    // getter
-    return this.get('bio') || "";
-  }.property('bio'),
-
-  truncatedAbout: function(key, value) {
-    // setter
-    if (arguments.length > 1) {
-      this.set('about', value.slice(0, 500));
-    }
-
-    // getter
-    return this.get('about') || "";
-  }.property('about')
+  }
 });
